+++ src/main/java/com/cells/gui/CellsGuiHandler.java
@@ -8,8 +8,10 @@ import net.minecraftforge.fml.common.network.IGuiHandler;
 
 import com.cells.blocks.importinterface.ContainerImportInterface;
 import com.cells.blocks.importinterface.ContainerMaxSlotSize;
+import com.cells.blocks.importinterface.ContainerPollingRate;
 import com.cells.blocks.importinterface.GuiImportInterface;
 import com.cells.blocks.importinterface.GuiMaxSlotSize;
+import com.cells.blocks.importinterface.GuiPollingRate;
 import com.cells.blocks.importinterface.TileImportInterface;
 
 
@@ -20,6 +22,7 @@ public class CellsGuiHandler implements IGuiHandler {
 
     public static final int GUI_IMPORT_INTERFACE = 0;
     public static final int GUI_MAX_SLOT_SIZE = 1;
+    public static final int GUI_POLLING_RATE = 2;
 
     @Override
     public Object getServerGuiElement(int id, EntityPlayer player, World world, int x, int y, int z) {
@@ -36,6 +39,12 @@ public class CellsGuiHandler implements IGuiHandler {
             case GUI_MAX_SLOT_SIZE:
                 if (tile instanceof TileImportInterface) {
                     return new ContainerMaxSlotSize(player.inventory, (TileImportInterface) tile);
+                }
+                break;
+
+            case GUI_POLLING_RATE:
+                if (tile instanceof TileImportInterface) {
+                    return new ContainerPollingRate(player.inventory, (TileImportInterface) tile);
                 }
                 break;
         }
@@ -58,6 +67,12 @@ public class CellsGuiHandler implements IGuiHandler {
             case GUI_MAX_SLOT_SIZE:
                 if (tile instanceof TileImportInterface) {
                     return new GuiMaxSlotSize(player.inventory, (TileImportInterface) tile);
+                }
+                break;
+
+            case GUI_POLLING_RATE:
+                if (tile instanceof TileImportInterface) {
+                    return new GuiPollingRate(player.inventory, (TileImportInterface) tile);
                 }
                 break;
         }