+++ src/main/java/com/cells/network/CellsNetworkHandler.java
@@ -3,6 +3,7 @@ package com.cells.network;
 import com.cells.Tags;
 import com.cells.network.packets.PacketOpenGui;
 import com.cells.network.packets.PacketSetMaxSlotSize;
+import com.cells.network.packets.PacketSetPollingRate;
 
 import net.minecraftforge.fml.common.network.NetworkRegistry;
 import net.minecraftforge.fml.common.network.simpleimpl.SimpleNetworkWrapper;
@@ -21,5 +22,6 @@ public class CellsNetworkHandler {
     public static void init() {
         INSTANCE.registerMessage(PacketSetMaxSlotSize.Handler.class, PacketSetMaxSlotSize.class, packetId++, Side.SERVER);
         INSTANCE.registerMessage(PacketOpenGui.Handler.class, PacketOpenGui.class, packetId++, Side.SERVER);
+        INSTANCE.registerMessage(PacketSetPollingRate.Handler.class, PacketSetPollingRate.class, packetId++, Side.SERVER);
     }
 }