+++ CHANGELOG.md
@@ -8,6 +8,35 @@ The format is based on Keep a Changelog and this project adheres to Semantic Ver
 - Semantic Versioning: https://semver.org/spec/v2.0.0.html
 
 
+## [0.3.8] - 2026-02-18
+### Fixed
+- Fix Import Interface not allowing items with the same ID and metadata but different NBT in filters, due NBT collision when checking for duplicates.
+
+
+## [0.3.7] - 2026-02-16
+### Fixed
+- Fix items with the same ID and metadata but different NBT, incorrectly merging in hyper-density cells (e.g., Thaumcraft vis crystals).
+
+
+## [0.3.6] - 2026-02-15
+### Fixed
+- Fix Import Interface's Polling Rate and Max Slot Size using short int instead of int, causing overflow and negative values when set above 32767.
+
+
+## [0.3.5] - 2026-02-14
+### Changed
+- Improve the tooltips to clarify some behaviors.
+- Tweak the Import Interface top and bottom.
+
+
+## [0.3.4] - 2026-02-14
+### Added
+- Add configurable polling rate to Import Interface, allowing users to set a fixed interval for importing items instead of relying on AE2's adaptive rates. This way, you can stock 100k items to send to the network every hour, for example.
+
+### Fixed
+- Fix some stale filter on world load for the Import Interface, which caused it to not import items until the filter was updated.
+
+
 ## [0.3.3] - 2026-02-13
 ### Added
 - Add upgrades support for Import Interface: Trash Unselected (voids items that don't match any filter) and Overflow (voids excess items that don't fit in the slot).